+++ packages/block-editor/src/components/inserter/index.native.js
@@ -15,6 +15,7 @@ import { compose, withPreferredColorScheme } from '@wordpress/compose';
 import { isUnmodifiedDefaultBlock } from '@wordpress/blocks';
 import {
 	Icon,
+	plusCircle,
 	plusCircleFilled,
 	insertAfter,
 	insertBefore,
@@ -64,7 +65,7 @@ export class Inserter extends Component {
 		const addBeforeOption = {
 			value: 'before',
 			label: __( 'Add Block Before' ),
-			icon: insertBefore,
+			icon: plusCircle,
 		};
 
 		const replaceCurrentOption = {
@@ -76,17 +77,17 @@ export class Inserter extends Component {
 		const addAfterOption = {
 			value: 'after',
 			label: __( 'Add Block After' ),
-			icon: insertAfter,
+			icon: plusCircle,
 		};
 
 		const addToBeginningOption = {
-			value: 'before',
+			value: 'start',
 			label: __( 'Add To Beginning' ),
 			icon: insertBefore,
 		};
 
 		const addToEndOption = {
-			value: 'after',
+			value: 'end',
 			label: __( 'Add To End' ),
 			icon: insertAfter,
 		};
@@ -95,12 +96,19 @@ export class Inserter extends Component {
 		if ( isAnyBlockSelected ) {
 			if ( isSelectedBlockReplaceable ) {
 				return [
+					addToBeginningOption,
 					addBeforeOption,
 					replaceCurrentOption,
 					addAfterOption,
+					addToEndOption,
 				];
 			}
-			return [ addBeforeOption, addAfterOption ];
+			return [
+				addToBeginningOption,
+				addBeforeOption,
+				addAfterOption,
+				addToEndOption,
+			];
 		}
 		return [ addToBeginningOption, addToEndOption ];
 	}
@@ -108,14 +116,22 @@ export class Inserter extends Component {
 	getInsertionIndex( insertionType ) {
 		const {
 			insertionIndexDefault,
+			insertionIndexStart,
 			insertionIndexBefore,
 			insertionIndexAfter,
+			insertionIndexEnd,
 		} = this.props;
+		if ( insertionType === 'start' ) {
+			return insertionIndexStart;
+		}
 		if ( insertionType === 'before' || insertionType === 'replace' ) {
 			return insertionIndexBefore;
 		}
 		if ( insertionType === 'after' ) {
 			return insertionIndexAfter;
+		}
+		if ( insertionType === 'end' ) {
+			return insertionIndexEnd;
 		}
 		return insertionIndexDefault;
 	}
@@ -335,19 +351,25 @@ export default compose( [
 			return endOfRootIndex;
 		}
 
+		const insertionIndexStart = 0;
+
 		const insertionIndexBefore = isAnyBlockSelected
 			? selectedBlockIndex
-			: 0;
+			: insertionIndexStart;
 
 		const insertionIndexAfter = isAnyBlockSelected
 			? selectedBlockIndex + 1
 			: endOfRootIndex;
+
+		const insertionIndexEnd = endOfRootIndex;
 
 		return {
 			destinationRootClientId,
 			insertionIndexDefault: getDefaultInsertionIndex(),
 			insertionIndexBefore,
 			insertionIndexAfter,
+			insertionIndexStart,
+			insertionIndexEnd,
 			isAnyBlockSelected,
 			isSelectedBlockReplaceable: isSelectedUnmodifiedDefaultBlock,
 		};