+++ packages/edit-site/src/store/test/actions.js
@@ -1,6 +1,7 @@
 /**
  * Internal dependencies
  */
+import registerEditSiteStore from '..';
 import {
 	toggleFeature,
 	setTemplate,
@@ -11,6 +12,8 @@ import {
 	showHomepage,
 	setHomeTemplateId,
 } from '../actions';
+
+registerEditSiteStore();
 
 describe( 'actions', () => {
 	describe( 'toggleFeature', () => {