+++ packages/edit-site/src/store/test/reducer.js
@@ -22,6 +22,9 @@ import {
 	setIsNavigationPanelOpened,
 	setIsInserterOpened,
 } from '../actions';
+import registerEditSiteStore from '..';
+
+registerEditSiteStore();
 
 describe( 'state', () => {
 	describe( 'preferences()', () => {